+++ sdk/keyvault/azcertificates/client.go
@@ -1051,10 +1051,6 @@ type UpdateCertificatePropertiesOptions struct {
 
 }
 
-func (u *UpdateCertificatePropertiesOptions) toGenerated() *generated.KeyVaultClientUpdateCertificateOptions {
-	return &generated.KeyVaultClientUpdateCertificateOptions{}
-}
-
 // UpdateCertificatePropertiesResponse contains response fields for Client.UpdateCertificateProperties
 type UpdateCertificatePropertiesResponse struct {
 	Certificate
@@ -1062,24 +1058,24 @@ type UpdateCertificatePropertiesResponse struct {
 
 // UpdateCertificateProperties applies the specified update on the given certificate; the only elements updated are the certificate's
 // attributes. This operation requires the certificates/update permission.
-func (c *Client) UpdateCertificateProperties(ctx context.Context, certificateName string, properties Properties, options *UpdateCertificatePropertiesOptions) (UpdateCertificatePropertiesResponse, error) {
-	if options == nil {
-		options = &UpdateCertificatePropertiesOptions{}
+func (c *Client) UpdateCertificateProperties(ctx context.Context, properties Properties, options *UpdateCertificatePropertiesOptions) (UpdateCertificatePropertiesResponse, error) {
+	name, version := "", ""
+	if properties.Name != nil {
+		name = *properties.Name
 	}
-	version := ""
 	if properties.Version != nil {
 		version = *properties.Version
 	}
 	resp, err := c.genClient.UpdateCertificate(
 		ctx,
 		c.vaultURL,
-		certificateName,
+		name,
 		version,
 		generated.CertificateUpdateParameters{
 			CertificateAttributes: properties.toGenerated(),
 			Tags:                  properties.Tags,
 		},
-		options.toGenerated(),
+		nil,
 	)
 	if err != nil {
 		return UpdateCertificatePropertiesResponse{}, err