+++ sdk/keyvault/azkeys/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Features Added
 * Added `NewCryptoClient()` to `azkeys.Client` to simplify access to the crypto client.
+* `UpdateKeyProperties()` can set a key's allowed operations
 
 ### Breaking Changes
 * Renamed methods which return `Pager[T]`:
@@ -20,6 +21,8 @@
 * Changed type of key `Tags` to `map[string]*string`
 * Changed type of `ListPropertiesOfKeyVersionsResponse.Keys` to `[]*KeyItem`
 * Changed type of `JSONWebKey.KeyOps` to `[]*Operation`
+* Moved `Key.ReleasePolicy` to `Key.Properties.ReleasePolicy`
+* `UpdateKeyProperties()` has a `Properties` parameter instead of a `Key` parameter
 
 ### Bugs Fixed
 * `ReleaseKey()` returns an error when no key version is specified